pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pptgen"
version = "0.1.0"
description = "Template-driven PowerPoint generation platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "python-pptx>=0.6",
]

[project.scripts]
pptgen = "pptgen.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pptgen"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
